song_backend/pipeline/requirements.txt
Flask==2.3.0
transformers==4.34.0
torch==2.0.0
Pillow==10.0.0
requests==2.31.0
numpy<2
openai==0.28.0